requirements.txt
Django==3.1.*
django-environ==0.4.*
psycopg2==2.8.*
djangorestframework==3.12.*
drf-yasg==1.21.*